requirements3.txt
-e git://github.com/closeio/cleancat.git#egg=cleancat-dev
mongoengine
flask-mongoengine
mimerender
protobuf==3.4.0
pycrypto==2.6
python-dateutil
sphinx
Flask>=0.9
Flask-Views
pymongo
flake8
six==1.11.0
-e git+ssh://git@github.com/closeio/zbase62.git@e13d2c748ccdb0cafe6465961a0c6a4111ee219f#egg=zbase62
